SIM-2858: Big refactor, mostly moving things around.
* Runtime now holds the API and is in ./runtime.py
* API now lives in ./api.py
* Bridge is MUCH simpler - most things moved to Runtime.
* Bridge now mostly focuses on startup and routing incoming messages while
  Runtime focuses on outgoing messages (and their ACKS).
* thistle_gulch/__init__.py is shared stuff inclding the router.
