+++ python/thistle_gulch/bridge.py
@@ -1,32 +1,16 @@
 import asyncio
-import json
-import logging
-import sys
-import traceback
-import uuid
-from datetime import datetime, timedelta
-from typing import Callable, Union, Dict, Type, Any, TypeVar, Optional, Awaitable
+from typing import Callable, Optional, Awaitable
 
-T = TypeVar('T')
-
-import cattrs
 import socketio
 from aiohttp import web
 from attr import define
 import fable_saga.server as saga_server
 
-import thistle_gulch
-from . import Runtime, Simulation, API
+from thistle_gulch.runtime import Runtime
+from . import GenericMessage, IncomingMessageRouter, Route, logger, parse_runtime_path_and_args
 from .data_models import PersonaContextObject
 
-logger = logging.getLogger(__name__)
-
-# module level converter to convert between objects and dicts.
-converter = cattrs.Converter(forbid_extra_keys=True)
-# Register a hook to convert datetime objects to and from isoformat strings.
-converter.register_unstructure_hook(datetime, lambda dt: dt.isoformat())
-converter.register_structure_hook(datetime, lambda dt: dt.fromisoformat(dt))
-
+# Set the saga_server converter to allow extra keys in for now.
 saga_server.converter.forbid_extra_keys = False
 
 
@@ -43,20 +27,6 @@ class TGConversationRequest(saga_server.ConversationRequest):
     context_obj: PersonaContextObject = None
 
 
-"""
-Sets up a server that can be used to generate actions for SAGA. Either HTTP or socketio can be used.
-"""
-
-
-@define(slots=True)
-class GenericMessage:
-    """ A generic message that is received."""
-    type: str
-    data: dict = {}
-    reference: str = None
-    error: str = None
-
-
 @define(slots=True)
 class BridgeConfig:
     host: str = 'localhost'
@@ -68,134 +38,14 @@ class BridgeConfig:
     conversation_endpoint = saga_server.ConversationServer()
 
 
-class Route:
-    def __init__(self, msg_type: str, msg_class: Type[T], process_function: Callable[[T], Awaitable[None]]):
-        self.msg_type = msg_type
-        self.msg_class = msg_class
-        self.process_function = process_function
-
-
-class IncomingMessageRouter:
-
-    def __init__(self):
-        self.routes: Dict[str, Route] = {}
-
-    def add_route(self, route: Route):
-        self.routes[route.msg_type] = route
-
-    async def handle_message(self, sid: str, message_str: str) -> Any:
-        logger.debug(f"[Message] received from {sid}: {message_str}")
-        try:
-            # Convert the message to a dictionary and validate that it has a type and that the type has a known route.
-            message = json.loads(message_str)
-            logger.info(
-                f"[Message] sid: {sid}, type: {message.get('type', 'unknown')}, ref: {message.get('reference', 'unknown')}")
-            msg_type = message.get('type')
-            if msg_type is None:
-                raise ValueError("Message is missing a type.")
-            if msg_type not in self.routes:
-                raise ValueError(f"Unknown message type: {msg_type}")
-
-            # Convert the message to the appropriate type and process it.
-            request_type = self.routes[message['type']].msg_class
-            request = converter.structure(message, request_type)
-            result = await self.routes[message['type']].process_function(request)
-
-            # If the result is None, then we don't send a response.
-            if result is None:
-                return
-
-            # Convert the result to a response and send it back.
-            response = converter.unstructure(result)
-            return response
-        except json.decoder.JSONDecodeError as e:
-            error = f"Error decoding JSON: {str(e)}"
-            ex = e
-        except cattrs.errors.ClassValidationError as e:
-            error = f"Error validating request: {json.dumps(cattrs.transform_error(e))}"
-            ex = e
-        except Exception as e:
-            error = f"Error processing request: {str(e)}"
-            ex = e
-        # If we get here, then there was an error.
-        logger.error(f"[Error] sid: {sid}, message: {message_str}, error: {error}, "
-                     f"stacktrace: {''.join(traceback.TracebackException.from_exception(ex).format())}")
-        response = saga_server.ErrorResponse(error=error)
-        output = converter.unstructure(response)
-        return output
-
-
 class RuntimeBridge:
 
     def __init__(self, config: BridgeConfig):
         self.config = config
-        self.runtime = None
-        self.simulation = Simulation()
         self.router = IncomingMessageRouter()
         self.on_ready: Optional[Callable[[RuntimeBridge], Awaitable[None]]] = None
         self.emit_lock = asyncio.Lock()
-        self.api = API(self)
-
-        # Add routes
-        async def on_ready_handler(ready_response: GenericMessage):
-            """Handler for the simulation-ready message."""
-            logger.info(f"[Simulation Ready] received..")
-            if self.on_ready is not None:
-                # pass the bridge instance to the on_ready callback so that it can send messages to the runtime.
-                logger.info(f"[Simulation Ready] Calling on_ready callback..")
-                await self.on_ready(self)
-            else:
-                logger.info(f"[Simulation Ready] No on_ready callback registered..")
-
-            logger.info(f"[Simulation Ready] Resuming simulation..")
-            await self.api.resume()
-
-        self.router.add_route(Route('simulation-ready', GenericMessage, on_ready_handler))
-
-        # Add some dummy routes for testing.
-        async def dummy_handler(request: GenericMessage):
-            print(request)
-            return None
-
-        async def dummy_handler_with_response(request: GenericMessage):
-            print(request)
-            return GenericMessage(type='heartbeat-ack', data={'ack': 'ack'}, reference=request.reference)
-
-        async def dummy_handler_independent_response(request: GenericMessage):
-            await asyncio.sleep(.2)
-
-            def test_callback(response: GenericMessage):
-                print(response)
-
-            await self.send_message('heartbeat-independent', {'ack': 'ack'}, test_callback)
-
-        self.router.add_route(Route('heartbeat', GenericMessage, dummy_handler_independent_response))
-        self.router.add_route(Route('heartbeat-ack', GenericMessage, dummy_handler_with_response))
-
-        # Validate the runtime path and create a runtime instance.
-        if self.config.runtime_path is not None:
-            split_args = self.config.runtime_path.split()
-            assert len(split_args) > 0, f"Error: Empty --runtime path"
-
-            runtime_arg_index = 1
-            runtime_path_str = split_args[0]
-            for runtime_arg_index in range(1, len(split_args)):
-                arg = split_args[runtime_arg_index]
-                # Detect the first runtime flag
-                if arg.startswith("-"):
-                    break
-                # Re-join the runtime path if it has spaces
-                else:
-                    runtime_path_str += " " + arg
-
-            # Validate runtime path
-            import pathlib
-            path = pathlib.Path(runtime_path_str)
-            assert path.exists(), f"Error: --runtime path not found: \"{runtime_path_str}\""
-            assert path.is_file(), f"Error: --runtime path is not a file: \"{runtime_path_str}\""
-
-            runtime_args = split_args[runtime_arg_index:]
-            self.runtime = Runtime(runtime_path_str, runtime_args)
+        self.runtime: Optional[Runtime] = None
 
         # Set up the async web server via aiohttp - this is the server that will handle the socketio connections.
         self.app = web.Application()
@@ -205,27 +55,46 @@ class RuntimeBridge:
         sio.attach(self.app)
         self.sio = sio
 
-        @sio.event
-        def connect(sid, _):
-            self.simulation.sim_id = sid
-            logger.info("Runtime [Connected]: " + sid)
+        # Validate the runtime path and create a runtime instance.
+        if self.config.runtime_path is not None:
+            runtime_exec, runtime_args = parse_runtime_path_and_args(self.config.runtime_path)
+
+            # Create the runtime instance.
+            # TODO: Refactor to allow multiple runtimes running at the same time, linking sim_id and sid to the runtime.
+            runtime = Runtime(sio, runtime_exec, runtime_args)
+            # Associate the runtime with the socketio server so it can send messages to the runtime.
+            runtime.sio = sio
+            self.runtime = runtime
+
+
+        ###
+        # BASIC SOCKETIO EVENT HANDLERS
+        ###
 
         @sio.event
-        def disconnect(sid):
-            self.simulation.sim_id = None
-            logger.info("Runtime [Disconnected]: " + sid)
+        async def connect(sid, _):
+            logger.info("[Socketio] Connected: " + sid)
+            # Associate the sid with the runtime via its on_connect method.
+            if self.runtime is None:
+                logger.warning("Runtime is connecting to a runtime that it didn't create. Creating a new runtime.")
+                self.runtime = Runtime(sio)
+            await self.runtime.on_connect(sid)
+
+        @sio.event
+        async def disconnect(sid):
+            logger.info("[Socketio] Disconnected: " + sid)
+            await self.runtime.on_disconnect(sid)
 
         # Set up the socketio event handlers. These are basically channels that the server listens on.
         # The first one is a catch-all for unhandled events.
         @sio.on('*')
         def catch_all(event, sid, *data):
             """Catch all unhandled events that have one message. (common)"""
-            logger.error(f"Unhandled event: {event} {sid} {data}")
+            logger.error(f"[Socketio] Unhandled event: {event} {sid} {data}")
 
-        # This is the main event that the server listens on when talking to the runtime.
-        @sio.on('messages')
-        async def on_incoming_messages(sid, message_str: str):
-            return await self.router.handle_message(sid, message_str)
+        ###
+        # STANDARD SAGA SERVER ENDPOINTS (LEGACY)
+        ###
 
         # These are the events that the server listens on when emulating the SAGA server. [Deprecated]
         @sio.on('generate-actions')
@@ -244,6 +113,31 @@ class RuntimeBridge:
             return await saga_server.generic_handler(message_str, TGConversationRequest,
                                                      self.config.conversation_endpoint.generate_conversation,
                                                      saga_server.ConversationResponse)
+
+        ###
+        # NEW RUNTIME BRIDGE ENDPOINTS (USES SINGLE STANDARD MESSAGES CHANNEL AND ROUTER)
+        ###
+        @sio.on('messages')
+        async def on_incoming_messages(sid, message_str: str):
+            return await self.router.handle_message(sid, message_str)
+
+        # Add routes
+        async def on_ready_handler(sid: str, ready_response: GenericMessage):
+            """Handler for the simulation-ready message."""
+            logger.info(f"[Simulation Ready] received..")
+
+            # Call the on_ready callback if it exists.
+            if self.on_ready is not None:
+                # pass the bridge instance to the on_ready callback so that it can send messages to the runtime.
+                logger.info(f"[Simulation Ready] Calling on_ready callback..")
+                await self.on_ready(self)
+            else:
+                logger.info(f"[Simulation Ready] No on_ready callback registered..")
+
+            logger.info(f"[Simulation Ready] Resuming simulation..")
+            await self.runtime.api.resume()
+
+        self.router.add_route(Route('simulation-ready', GenericMessage, on_ready_handler))
 
     def run(self):
         print("""
@@ -272,87 +166,6 @@ class RuntimeBridge:
             self.runtime.terminate()
             logger.info('Runtime [Stopped]')
 
-    async def send_message(self, msg_type, data, timeout=5) -> GenericMessage:
-        """
-               Send a request to the runtime.
-               :param msg_type: type of message.
-               :param data: data to send with the message.
-               :param callback: [Optional] function to call when the server responds.
-               """
-
-        # Create a Future object that we will use to wait for the callback response
-        future = asyncio.Future()
-
-        # Define a callback function that will be called when the emit() receives a response
-        async def callback(response_type: str, response_data: str):
-            self.emit_lock.release()
-            # If the future is already done (cancelled), then we don't need to do anything.
-            if future.done():
-                return None
-            # Set the future result, which will unblock the send_message function.
-            future.set_result((response_type, response_data))
-
-        if self.simulation.sim_id is None:
-            err = f'[Message Request] Error: simulation client id is None.'
-            logging.error(err)
-            raise ValueError(err)
-
-        # Create a reference for the message so that we can match the response to the request.
-        reference = uuid.uuid4().hex
-        # Create a GenericMessage object and serialize it to a string so that it can be sent to the Runtime.
-        request_msg = GenericMessage(type=msg_type, data=data, reference=reference)
-        serialized_message = json.dumps(converter.unstructure(request_msg))
-
-        # Acquire the emit lock to prevent multiple emits from happening at the same time.
-        await self.emit_lock.acquire()
-        try:
-            await self.sio.emit('messages', (request_msg.type, serialized_message), to=self.simulation.sim_id,
-                                callback=callback)
-        except Exception as e:
-            err = f"[Message Request] Error sending message: {str(e)}"
-            logging.error(err)
-            # Release the emit lock if an error occurs.
-            self.emit_lock.release()
-            raise e
-
-        # Wait for the future to have a result, with timeout
-        try:
-            # Wait for the Future object to have a result, with timeout
-            response_type, response_data = await asyncio.wait_for(future, timeout)
-        except asyncio.TimeoutError as e:
-            # Handle the case where the future times out
-            err = f"[Message Response] Response timed out after {timeout} seconds"
-            logging.error(err)
-            raise e
-        except Exception as e:
-            # Handle other exceptions that could occur
-            err = f"[Message Response] Response exception: {str(e)}"
-            logging.error(err)
-            raise e
-
-        # Parse the response_data and set the future result.
-        try:
-            data = json.loads(response_data)
-        except json.JSONDecodeError as e:
-            err = f'[Message Response] {response_type} Error decoding JSON: {str(e)}'
-            logging.error(err)
-            raise e
-
-        if data.get('error'):
-            err = f'[Message Response] {response_type} Runtime Error: {data.get("error")}'
-            logging.error(err)
-            raise ValueError(err)
-
-        # Validate that the response_data is a dictionary.
-        if not isinstance(data, dict):
-            err = f'[Message Response] Error: data not a dictionary.'
-            logging.error(err)
-            raise ValueError(err)
-
-        # Convert to a GenericMessage.
-        response = GenericMessage(type=response_type, data=data, reference=reference)
-        return response
-
 
 def main(auto_run=True) -> RuntimeBridge:
     dummy_config = BridgeConfig()
@@ -373,9 +186,3 @@ def main(auto_run=True) -> RuntimeBridge:
     if auto_run:
         bridge.run()
     return bridge
-
-
-if __name__ == '__main__':
-    logging.basicConfig(level=logging.INFO, stream=sys.stdout,
-                        format='<%(levelname)s> %(asctime)s - %(name)s - %(pathname)s:%(lineno)d\n    %(message)s')
-    main()
